📄 Add playground, create .gitignore and README.md
A visitor should have a view of the general information such as availability and used frameworks. Therefore I created a README.md file to include: a link to the article, the minimum required Xcode version, the frameworks, and a screenshot.

The repository doesn't include the demo project. Therefore, I added the corresponding Xcode Playground.

Some files are unuseful to the repository, and Git should ignore them.
Therefore, I created the .gitignore file with the files, directories and extensions to ignore.